Partial completed feature, push ( see notes )
I needed to move the work from one laptop to another, so work around issues with GoLand IDE etc..
Plus it's always useful to be able to move work around using Git. This also shares the work to Github so
I can work on it from where-ever I like.
In regard to this feature, the view templates need more work, the images model is not working correctly. The facilities, and owning user relational models
are almost done, but the view templates need to handle the data elegently.